pyproject.toml
[project]
name = "youtube-chitra-janak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "playwright>=1.57.0",
    "pyyaml>=6.0.3",
]
